fix the way of placing points in 3D with offset and zoom value included

+++ fdf.h
@@ -6,7 +6,7 @@
 /*   By: algasnie <algasnie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/27 13:47:27 by algasnie          #+#    #+#             */
-/*   Updated: 2025/12/03 14:54:26 by algasnie         ###   ########.fr       */
+/*   Updated: 2025/12/03 15:29:34 by algasnie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,9 +43,9 @@ typedef struct s_view
 
 typedef struct s_point_proj
 {
-	double	x;
-	double	y;
-	double	z;
+	int	x;
+	int	y;
+	int	z;
 	int	color;
 } t_point_proj;
